service_2/requirements.txt
aws_cdk.core
aws_cdk.aws_kinesis
aws_cdk.aws_lambda
aws_cdk.aws_ec2
cfn-lint
